Move ringbuffer to its own console
